Moving test to testing_test to respect style guide
